[buteo-sync-plugin-carddav] Improve delta determination logic. Contributes to JB#1665
Some CardDAV servers return the status for a response outside of the
propstat element.  Previously, the code assumed that the status would
be contained within the propstat element, this commit ensures that
we handle that status information.

This commit also fixes a bug in the upsync-delta-determination code
which would misreport whether the local contact had differences
to the remote contact in most cases, resulting in either a spurious
upsync or a missed upsync.

Contributes to JB#1665