[buteo-sync-plugin-carddav] Detect non-addressbook resources. Contributes to MER#1657
This commit adds a check in our handling for resources during
addressbook information parsing, by ensuring that we require that
a resource be an addressbook collection resource in order to be
treated as one.  Previously, the code only checked that the
status code associated with the resource was 200 OK, due to an
omission of a variable check.

Contributes to MER#1657