[buteo-sync-plugin-carddav] Allow addressbook path as home set url. Contributes to JB#38586
In some cases, the endpoint which some providers (e.g. Kerio) provide
to users will not be an addressbook path but instead a home-set-url.
This commit ensures that if addressbook discovery fails at the given
path, that we try to treat the path as a home-set-url, which then
allows discovery of the addressbook paths.

It also ensures that we compare the ctag or sync-token for a given
addressbook path correctly with the one stored from the previous sync.

Contributes to MER#1751 and JB#38586