[buteo-sync-plugin-carddav] Allow sync of non .vcf resources. Contributes to MER#1863
Previously, we only synced resources returned by the server in
response to an etags request, if those resources ended in ".vcf".

This commit modifies the behaviour so that we assume that any resource
which has a suffix which is not ".vcf" (e.g., ".ics" or ".eml") should
be ignored, but other resources without a suffix should be synced.

This allows us to sync in cases where the remote resource is found at
a path like "/carddav/Addressbooks/personal/johnsmith" rather than
a path like "/carddav/Addressbooks/personal/johnsmith.vcf".

Contributes to MER#1863