[buteo-sync-plugin-carddav] Prefer definite addressbook resources. Contributes to JB#43847
Some services (e.g. Memotoo) do not produce output which conforms
to the RFC, specifically they don't tag their addressbook resources
with the appropriate resource-type tag.

Previously, we worked around that case by assuming that a resource
which otherwise meet the requirements must be an addressbook resource.

This commit fixes the behaviour so that we only assume that such
a resource is an addressbook resource if NO OTHER resources are
explicitly tagged with the appropriate resource-type tag.

This fixes upsync with e.g. Yandex.

Contributes to JB#43847