[buteo-sync-plugin-carddav] Allow redirect with different scheme. Contributes to JB#53434
Previously, if the server redirected a request from:
    https://dav.server.tld/
to:
    http://dav.server.tld/
or vice versa, we wrongly detected the redirect as spurious due to
the paths matching.

This commit allows the redirect due to the different scheme.
It also allows redirecting to a different port.

However, if the host is different, the redirect is blocked,
as that is unsafe.