[buteo-sync-plugin-carddav] Resolve relative redirects before validation
The validation code for redirects always assumes the redirect is an
absolute URL. When the redirect is not an absolute URL, the hostname
of the unresolved redirect URL is empty, while the hostname of the
original URL isn't.  Hence, hostChanged will always be true and the
validation code will throw an error, while it shouldn't.

This commit fixes the following issue reported on the Sailfish Community
Forums: https://forum.sailfishos.org/t/11033